R6 Class Representing a User

The [FPL$get_user()] method returns an object of class `User`.

Public fields

user

A `list` of user details.

history

A `list` of `data.frame`s containing the historic gameweek performance data for the user.

cup

A `list` of data related to the user's cup performance.

picks

A `list` of the user's picks information for each gameweek.

transfers

A `list` of the user's weekly transfers.

Methods

Public methods

User$new()

User$get_gameweek_history()

User$get_season_history()

User$get_chips_history()

User$get_picks()

User$get_cup_status()

User$get_cup_matches()

User$get_active_chips()

User$get_automatic_substitutions()

User$get_user_history()

User$get_transfers()

User$print()

User$clone()

Method new()

Instantiate the `User` class.

Usage

User$new(user)

Arguments

user

`list`. Data for a single FPL user.

Returns

An R6 `User` object.

Examples

fpl <- FPL$new()
user <- fpl$get_user(555690)

Method get_gameweek_history()

Get the gameweek history for the user.

Usage

User$get_gameweek_history(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A `data.frame`.

Examples

user$get_gameweek_history()

Method get_season_history()

Get the seasonal history for the user.

Usage

User$get_season_history()

Returns

A `data.frame`.

Examples

user$get_season_history()

Method get_chips_history()

Get the chip history for the user.

Usage

User$get_chips_history(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A `data.frame`.

Examples

user$get_gameweek_history()

Method get_picks()

Get the user's picks for the chosen `gameweek`(s).

Usage

User$get_picks(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek picks to return (default: `NULL` returns all gameweeks).

Returns

A `list` containing the user's picks details for the chosen gameweek(s).

Examples

user$get_picks()

Method get_cup_status()

Get the user's cup status.

Usage

User$get_cup_status()

Returns

A `list` of cup status information.

Examples

user$get_cup_status()

Method get_cup_matches()

Get the user's cup matches.

Usage

User$get_cup_matches(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A `list` of cup match information.

Examples

user$get_cup_matches()

Method get_active_chips()

Get the user's active chip for each gameweek or the active chip of the chosen `gameweek`(s).

Usage

User$get_active_chips(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A vector of active chips.

Examples

user$get_active_chips()

Method get_automatic_substitutions()

Get a `list` of automatic substitutions for the chosen `gameweek`(s).

Usage

User$get_automatic_substitutions(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A vector of active chips.

Examples

user$get_automatic_substitutions()

Method get_user_history()

Get a `list` of the user's `gameweek` history.

Usage

User$get_user_history(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A `list` of `data.frame`s of the user's gameweek history.

Examples

user$get_user_history()

Method get_transfers()

Get a `list` of all the user's transfers, or a `list` of transfers made in the given `gameweek`(s).

Usage

User$get_transfers(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A `list` of the user's transfers.

Examples

user$get_transfers()

Method print()

Pretty printing of the `User`.

Usage

User$print(...)

Arguments

...

Not used.

Method clone()

The objects of this class are cloneable with this method.

Usage

User$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `User$new`
## ------------------------------------------------

fpl <- FPL$new()
user <- fpl$get_user(555690)

## ------------------------------------------------
## Method `User$get_gameweek_history`
## ------------------------------------------------

user$get_gameweek_history()

#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     1     67           67 1055000   1064199      1055000   30  1000
#> 2     2     50          117 4317434   4322575      1989996   30  1000
#> 3     3     43          156 3460040   3461869      2400783    2  1003
#> 4     4     67          223 1047802   1050883      1362049   12  1009
#> 5     5     77          300 1103908   1107719       845473    0  1010
#> 6     6     44          344 4401261   4407865      1087313    0  1018
#> 7     7     62          406 1698345   1703247       852783    0  1023
#> 8     8     59          461 3060497   3061774      1052450    0  1027
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               2
#> 2               0                    0               0
#> 3               3                    4               0
#> 4               1                    0               9
#> 5               0                    0              14
#> 6               0                    0               6
#> 7               0                    0               1
#> 8               3                    4               8

## ------------------------------------------------
## Method `User$get_season_history`
## ------------------------------------------------

user$get_season_history()

#>   season_name total_points   rank
#> 1     2018/19         2198 278647
#> 2     2019/20         2205 273955

## ------------------------------------------------
## Method `User$get_chips_history`
## ------------------------------------------------

user$get_gameweek_history()

#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     1     67           67 1055000   1064199      1055000   30  1000
#> 2     2     50          117 4317434   4322575      1989996   30  1000
#> 3     3     43          156 3460040   3461869      2400783    2  1003
#> 4     4     67          223 1047802   1050883      1362049   12  1009
#> 5     5     77          300 1103908   1107719       845473    0  1010
#> 6     6     44          344 4401261   4407865      1087313    0  1018
#> 7     7     62          406 1698345   1703247       852783    0  1023
#> 8     8     59          461 3060497   3061774      1052450    0  1027
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               2
#> 2               0                    0               0
#> 3               3                    4               0
#> 4               1                    0               9
#> 5               0                    0              14
#> 6               0                    0               6
#> 7               0                    0               1
#> 8               3                    4               8

## ------------------------------------------------
## Method `User$get_picks`
## ------------------------------------------------

user$get_picks()

#> [[1]]
#> [[1]]$active_chip
#> NULL
#> 
#> [[1]]$automatic_subs
#> list()
#> 
#> [[1]]$entry_history
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     1     67           67 1055000   1064199      1055000   30  1000
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               2
#> 
#> [[1]]$picks
#>    element position multiplier is_captain is_vice_captain
#> 1      363        1          1      FALSE           FALSE
#> 2      259        2          1      FALSE           FALSE
#> 3      467        3          1      FALSE           FALSE
#> 4      457        4          1      FALSE           FALSE
#> 5      254        5          2       TRUE           FALSE
#> 6      235        6          1      FALSE           FALSE
#> 7      338        7          1      FALSE           FALSE
#> 8        4        8          1      FALSE            TRUE
#> 9      184        9          1      FALSE           FALSE
#> 10     377       10          1      FALSE           FALSE
#> 11     117       11          1      FALSE           FALSE
#> 12      35       12          0      FALSE           FALSE
#> 13     196       13          0      FALSE           FALSE
#> 14      76       14          0      FALSE           FALSE
#> 15      97       15          0      FALSE           FALSE
#> 
#> 
#> [[2]]
#> [[2]]$active_chip
#> NULL
#> 
#> [[2]]$automatic_subs
#> list()
#> 
#> [[2]]$entry_history
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     2     50          117 4317434   4322575      1989996   30  1000
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               0
#> 
#> [[2]]$picks
#>    element position multiplier is_captain is_vice_captain
#> 1      363        1          1      FALSE           FALSE
#> 2      259        2          1      FALSE           FALSE
#> 3       97        3          1      FALSE           FALSE
#> 4      457        4          1      FALSE           FALSE
#> 5      254        5          2       TRUE           FALSE
#> 6      235        6          1      FALSE           FALSE
#> 7      338        7          1      FALSE           FALSE
#> 8        4        8          1      FALSE            TRUE
#> 9      184        9          1      FALSE           FALSE
#> 10     377       10          1      FALSE           FALSE
#> 11     117       11          1      FALSE           FALSE
#> 12      35       12          0      FALSE           FALSE
#> 13      76       13          0      FALSE           FALSE
#> 14     196       14          0      FALSE           FALSE
#> 15     467       15          0      FALSE           FALSE
#> 
#> 
#> [[3]]
#> [[3]]$active_chip
#> NULL
#> 
#> [[3]]$automatic_subs
#> list()
#> 
#> [[3]]$entry_history
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     3     43          156 3460040   3461869      2400783    2  1003
#>   event_transfers event_transfers_cost points_on_bench
#> 1               3                    4               0
#> 
#> [[3]]$picks
#>    element position multiplier is_captain is_vice_captain
#> 1      363        1          1      FALSE           FALSE
#> 2      259        2          1      FALSE           FALSE
#> 3       97        3          1      FALSE           FALSE
#> 4      457        4          1      FALSE           FALSE
#> 5      254        5          2       TRUE           FALSE
#> 6      235        6          1      FALSE           FALSE
#> 7      508        7          1      FALSE           FALSE
#> 8      272        8          1      FALSE            TRUE
#> 9      184        9          1      FALSE           FALSE
#> 10     164       10          1      FALSE           FALSE
#> 11     117       11          1      FALSE           FALSE
#> 12      35       12          0      FALSE           FALSE
#> 13     467       13          0      FALSE           FALSE
#> 14     196       14          0      FALSE           FALSE
#> 15      76       15          0      FALSE           FALSE
#> 
#> 
#> [[4]]
#> [[4]]$active_chip
#> NULL
#> 
#> [[4]]$automatic_subs
#> list()
#> 
#> [[4]]$entry_history
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     4     67          223 1047802   1050883      1362049   12  1009
#>   event_transfers event_transfers_cost points_on_bench
#> 1               1                    0               9
#> 
#> [[4]]$picks
#>    element position multiplier is_captain is_vice_captain
#> 1      363        1          1      FALSE           FALSE
#> 2      259        2          1      FALSE           FALSE
#> 3       97        3          1      FALSE           FALSE
#> 4      457        4          1      FALSE           FALSE
#> 5      254        5          2       TRUE           FALSE
#> 6      235        6          1      FALSE           FALSE
#> 7      508        7          1      FALSE           FALSE
#> 8      272        8          1      FALSE            TRUE
#> 9      184        9          1      FALSE           FALSE
#> 10     164       10          1      FALSE           FALSE
#> 11     460       11          1      FALSE           FALSE
#> 12      35       12          0      FALSE           FALSE
#> 13     467       13          0      FALSE           FALSE
#> 14      76       14          0      FALSE           FALSE
#> 15     196       15          0      FALSE           FALSE
#> 
#> 
#> [[5]]
#> [[5]]$active_chip
#> [1] "wildcard"
#> 
#> [[5]]$automatic_subs
#> list()
#> 
#> [[5]]$entry_history
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     5     77          300 1103908   1107719       845473    0  1010
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0              14
#> 
#> [[5]]$picks
#>    element position multiplier is_captain is_vice_captain
#> 1      363        1          1      FALSE           FALSE
#> 2      259        2          1      FALSE           FALSE
#> 3       97        3          1      FALSE           FALSE
#> 4       46        4          1      FALSE           FALSE
#> 5      254        5          2       TRUE           FALSE
#> 6       37        6          1      FALSE           FALSE
#> 7      508        7          1      FALSE           FALSE
#> 8      390        8          1      FALSE            TRUE
#> 9      388        9          1      FALSE           FALSE
#> 10     164       10          1      FALSE           FALSE
#> 11     460       11          1      FALSE           FALSE
#> 12      56       12          0      FALSE           FALSE
#> 13      76       13          0      FALSE           FALSE
#> 14     146       14          0      FALSE           FALSE
#> 15     470       15          0      FALSE           FALSE
#> 
#> 
#> [[6]]
#> [[6]]$active_chip
#> NULL
#> 
#> [[6]]$automatic_subs
#> list()
#> 
#> [[6]]$entry_history
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     6     44          344 4401261   4407865      1087313    0  1018
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               6
#> 
#> [[6]]$picks
#>    element position multiplier is_captain is_vice_captain
#> 1      363        1          1      FALSE           FALSE
#> 2      259        2          1      FALSE           FALSE
#> 3       46        3          1      FALSE           FALSE
#> 4      470        4          1      FALSE           FALSE
#> 5      254        5          2       TRUE           FALSE
#> 6       37        6          1      FALSE           FALSE
#> 7      508        7          1      FALSE           FALSE
#> 8      390        8          1      FALSE            TRUE
#> 9      460        9          1      FALSE           FALSE
#> 10     164       10          1      FALSE           FALSE
#> 11     388       11          1      FALSE           FALSE
#> 12      56       12          0      FALSE           FALSE
#> 13      76       13          0      FALSE           FALSE
#> 14      97       14          0      FALSE           FALSE
#> 15     146       15          0      FALSE           FALSE
#> 
#> 
#> [[7]]
#> [[7]]$active_chip
#> NULL
#> 
#> [[7]]$automatic_subs
#>    entry element_in element_out event
#> 1 555690         76         508     7
#> 
#> [[7]]$entry_history
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     7     62          406 1698345   1703247       852783    0  1023
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               1
#> 
#> [[7]]$picks
#>    element position multiplier is_captain is_vice_captain
#> 1      363        1          1      FALSE           FALSE
#> 2      259        2          1      FALSE           FALSE
#> 3      470        3          1      FALSE           FALSE
#> 4       46        4          1      FALSE           FALSE
#> 5       76        5          1      FALSE           FALSE
#> 6      254        6          2       TRUE           FALSE
#> 7      390        7          1      FALSE            TRUE
#> 8       37        8          1      FALSE           FALSE
#> 9      460        9          1      FALSE           FALSE
#> 10     388       10          1      FALSE           FALSE
#> 11     164       11          1      FALSE           FALSE
#> 12      56       12          0      FALSE           FALSE
#> 13     508       13          0      FALSE           FALSE
#> 14      97       14          0      FALSE           FALSE
#> 15     146       15          0      FALSE           FALSE
#> 
#> 
#> [[8]]
#> [[8]]$active_chip
#> NULL
#> 
#> [[8]]$automatic_subs
#> list()
#> 
#> [[8]]$entry_history
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     8     59          461 3060497   3061774      1052450    0  1027
#>   event_transfers event_transfers_cost points_on_bench
#> 1               3                    4               8
#> 
#> [[8]]$picks
#>    element position multiplier is_captain is_vice_captain
#> 1      363        1          1      FALSE           FALSE
#> 2      110        2          1      FALSE           FALSE
#> 3      470        3          1      FALSE           FALSE
#> 4       97        4          1      FALSE           FALSE
#> 5      254        5          1      FALSE           FALSE
#> 6      272        6          1      FALSE           FALSE
#> 7      508        7          1      FALSE           FALSE
#> 8      390        8          1      FALSE            TRUE
#> 9      506        9          1      FALSE           FALSE
#> 10     164       10          1      FALSE           FALSE
#> 11     388       11          2       TRUE           FALSE
#> 12      56       12          0      FALSE           FALSE
#> 13      76       13          0      FALSE           FALSE
#> 14      46       14          0      FALSE           FALSE
#> 15     146       15          0      FALSE           FALSE
#> 
#> 

## ------------------------------------------------
## Method `User$get_cup_status`
## ------------------------------------------------

user$get_cup_status()

#> $qualification_event
#> [1] 16
#> 
#> $qualification_numbers
#> [1] 4194304
#> 
#> $qualification_rank
#> NULL
#> 
#> $qualification_state
#> [1] "NOT_STARTED"
#> 

## ------------------------------------------------
## Method `User$get_cup_matches`
## ------------------------------------------------

user$get_cup_matches()

#> list()

## ------------------------------------------------
## Method `User$get_active_chips`
## ------------------------------------------------

user$get_active_chips()

#> [[1]]
#> NULL
#> 
#> [[2]]
#> NULL
#> 
#> [[3]]
#> NULL
#> 
#> [[4]]
#> NULL
#> 
#> [[5]]
#> [1] "wildcard"
#> 
#> [[6]]
#> NULL
#> 
#> [[7]]
#> NULL
#> 
#> [[8]]
#> NULL
#> 

## ------------------------------------------------
## Method `User$get_automatic_substitutions`
## ------------------------------------------------

user$get_automatic_substitutions()

#> [[1]]
#> list()
#> 
#> [[2]]
#> list()
#> 
#> [[3]]
#> list()
#> 
#> [[4]]
#> list()
#> 
#> [[5]]
#> list()
#> 
#> [[6]]
#> list()
#> 
#> [[7]]
#>    entry element_in element_out event
#> 1 555690         76         508     7
#> 
#> [[8]]
#> list()
#> 

## ------------------------------------------------
## Method `User$get_user_history`
## ------------------------------------------------

user$get_user_history()

#> [[1]]
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     1     67           67 1055000   1064199      1055000   30  1000
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               2
#> 
#> [[2]]
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     2     50          117 4317434   4322575      1989996   30  1000
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               0
#> 
#> [[3]]
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     3     43          156 3460040   3461869      2400783    2  1003
#>   event_transfers event_transfers_cost points_on_bench
#> 1               3                    4               0
#> 
#> [[4]]
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     4     67          223 1047802   1050883      1362049   12  1009
#>   event_transfers event_transfers_cost points_on_bench
#> 1               1                    0               9
#> 
#> [[5]]
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     5     77          300 1103908   1107719       845473    0  1010
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0              14
#> 
#> [[6]]
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     6     44          344 4401261   4407865      1087313    0  1018
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               6
#> 
#> [[7]]
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     7     62          406 1698345   1703247       852783    0  1023
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               1
#> 
#> [[8]]
#>   event points total_points    rank rank_sort overall_rank bank value
#> 1     8     59          461 3060497   3061774      1052450    0  1027
#>   event_transfers event_transfers_cost points_on_bench
#> 1               3                    4               8
#> 

## ------------------------------------------------
## Method `User$get_transfers`
## ------------------------------------------------

user$get_transfers()

#>    element_in element_in_cost element_out element_out_cost  entry event
#> 1         506              65         460               85 555690     8
#> 2         272             115          37               72 555690     8
#> 3         110              51         259               74 555690     8
#> 4          46              46         457               59 555690     5
#> 5         388             106         184               59 555690     5
#> 6          37              71         235               70 555690     5
#> 7         146              41         467               40 555690     5
#> 8         390              91         272              116 555690     5
#> 9         470              40         196               39 555690     5
#> 10         56              40          35               40 555690     5
#> 11        460              85         117               95 555690     4
#> 12        508              75         338               55 555690     3
#> 13        272             116           4              120 555690     3
#> 14        164              72         377               60 555690     3
#>                           time
#> 1  2020-10-31T22:56:07.925243Z
#> 2  2020-10-31T22:56:07.924440Z
#> 3  2020-10-31T22:56:07.912237Z
#> 4  2020-10-16T16:06:35.465945Z
#> 5  2020-10-16T16:06:35.465582Z
#> 6  2020-10-16T16:06:35.465273Z
#> 7  2020-10-16T16:06:35.464957Z
#> 8  2020-10-16T16:06:35.464615Z
#> 9  2020-10-16T16:06:35.464298Z
#> 10 2020-10-16T16:06:35.463894Z
#> 11 2020-09-26T22:30:20.905715Z
#> 12 2020-09-23T21:55:44.080243Z
#> 13 2020-09-22T07:34:09.702596Z
#> 14 2020-09-20T16:28:45.550741Z