Filter player elements from the web

Filter player elements from the web

Usage

filter_players(data, get_photos = F, suppress = T)

Arguments

data

list from api

get_photos

optional boolean: If TRUE, will gather photos from FPL website

suppress

option boolean: If FALSE, will print out print messages and player photos not found.

Value

returns data frame with additional columns, player photo paths, and team abbreviations

Examples

if (FALSE) { # \dontrun{
data_list <- get_premier_stats()
players <- filter_players(data_list)
} # }