Get the current the stats from the fantasy API

Get the current the stats from the fantasy API

Usage

get_premier_stats()

Value

data frame with all players and variables

Examples

# Get the stats from online
weekly_stats <- get_premier_stats()
head(weekly_stats) # print stats
#> $chips
#>   id     name number start_event stop_event chip_type overrides.squad_squadsize
#> 1  1 wildcard      1           2         19  transfer                        NA
#> 2  2 wildcard      1          20         38  transfer                        NA
#> 3  3  freehit      1           2         38  transfer                        NA
#> 4  4   bboost      1           1         38      team                        NA
#> 5  5      3xc      1           1         38      team                        NA
#> 6  6  manager      1          24         38  transfer                        16
#>                                         overrides.element_types
#> 1                                                          NULL
#> 2                                                          NULL
#> 3                                                          NULL
#> 4                                                          NULL
#> 5                                                          NULL
#> 6 5, Managers, MNG, Manager, MNG, 1, NA, NA, 0, 0, TRUE, 16, 20
#>   overrides.pick_multiplier
#> 1                        NA
#> 2                        NA
#> 3                        NA
#> 4                        NA
#> 5                        NA
#> 6                        NA
#> 
#> $events
#>    id        name        deadline_time release_time average_entry_score
#> 1   1  Gameweek 1 2024-08-16T17:30:00Z           NA                  57
#> 2   2  Gameweek 2 2024-08-24T10:00:00Z           NA                  69
#> 3   3  Gameweek 3 2024-08-31T10:00:00Z           NA                  64
#> 4   4  Gameweek 4 2024-09-14T10:00:00Z           NA                  51
#> 5   5  Gameweek 5 2024-09-21T10:00:00Z           NA                  58
#> 6   6  Gameweek 6 2024-09-28T10:00:00Z           NA                  50
#> 7   7  Gameweek 7 2024-10-05T10:00:00Z           NA                  46
#> 8   8  Gameweek 8 2024-10-19T10:00:00Z           NA                  36
#> 9   9  Gameweek 9 2024-10-25T17:30:00Z           NA                  54
#> 10 10 Gameweek 10 2024-11-02T11:00:00Z           NA                  39
#> 11 11 Gameweek 11 2024-11-09T13:30:00Z           NA                  49
#> 12 12 Gameweek 12 2024-11-23T11:00:00Z           NA                  49
#> 13 13 Gameweek 13 2024-11-29T18:30:00Z           NA                  60
#> 14 14 Gameweek 14 2024-12-03T18:00:00Z           NA                  58
#> 15 15 Gameweek 15 2024-12-07T11:00:00Z           NA                  50
#> 16 16 Gameweek 16 2024-12-14T13:30:00Z           NA                  47
#> 17 17 Gameweek 17 2024-12-21T11:00:00Z           NA                  60
#> 18 18 Gameweek 18 2024-12-26T11:00:00Z           NA                  51
#> 19 19 Gameweek 19 2024-12-29T13:00:00Z           NA                  66
#> 20 20 Gameweek 20 2025-01-04T11:00:00Z           NA                  60
#> 21 21 Gameweek 21 2025-01-14T18:00:00Z           NA                  55
#> 22 22 Gameweek 22 2025-01-18T11:00:00Z           NA                  46
#> 23 23 Gameweek 23 2025-01-25T13:30:00Z           NA                   0
#> 24 24 Gameweek 24 2025-02-01T11:00:00Z           NA                   0
#> 25 25 Gameweek 25 2025-02-14T18:30:00Z           NA                   0
#> 26 26 Gameweek 26 2025-02-21T18:30:00Z           NA                   0
#> 27 27 Gameweek 27 2025-02-25T18:00:00Z           NA                   0
#> 28 28 Gameweek 28 2025-03-08T13:30:00Z           NA                   0
#> 29 29 Gameweek 29 2025-03-15T13:30:00Z           NA                   0
#> 30 30 Gameweek 30 2025-04-01T17:15:00Z           NA                   0
#> 31 31 Gameweek 31 2025-04-05T12:30:00Z           NA                   0
#> 32 32 Gameweek 32 2025-04-12T12:30:00Z           NA                   0
#> 33 33 Gameweek 33 2025-04-19T12:30:00Z           NA                   0
#> 34 34 Gameweek 34 2025-04-26T12:30:00Z           NA                   0
#> 35 35 Gameweek 35 2025-05-03T12:30:00Z           NA                   0
#> 36 36 Gameweek 36 2025-05-10T12:30:00Z           NA                   0
#> 37 37 Gameweek 37 2025-05-18T12:30:00Z           NA                   0
#> 38 38 Gameweek 38 2025-05-25T13:30:00Z           NA                   0
#>    finished data_checked highest_scoring_entry deadline_time_epoch
#> 1      TRUE         TRUE               3546234          1723829400
#> 2      TRUE         TRUE               9442126          1724493600
#> 3      TRUE         TRUE                430195          1725098400
#> 4      TRUE         TRUE               3560750          1726308000
#> 5      TRUE         TRUE               1773336          1726912800
#> 6      TRUE         TRUE               7479996          1727517600
#> 7      TRUE         TRUE                759474          1728122400
#> 8      TRUE         TRUE               5482232          1729332000
#> 9      TRUE         TRUE               2039494          1729877400
#> 10     TRUE         TRUE               1689659          1730545200
#> 11     TRUE         TRUE              10668888          1731159000
#> 12     TRUE         TRUE               7440806          1732359600
#> 13     TRUE         TRUE               1929511          1732905000
#> 14     TRUE         TRUE               2964307          1733248800
#> 15     TRUE         TRUE               7399737          1733569200
#> 16     TRUE         TRUE               5142221          1734183000
#> 17     TRUE         TRUE               4917384          1734778800
#> 18     TRUE         TRUE               2701866          1735210800
#> 19     TRUE         TRUE               3191421          1735477200
#> 20     TRUE         TRUE               1247541          1735988400
#> 21     TRUE         TRUE                434390          1736877600
#> 22     TRUE         TRUE               2700271          1737198000
#> 23    FALSE        FALSE                    NA          1737811800
#> 24    FALSE        FALSE                    NA          1738407600
#> 25    FALSE        FALSE                    NA          1739557800
#> 26    FALSE        FALSE                    NA          1740162600
#> 27    FALSE        FALSE                    NA          1740506400
#> 28    FALSE        FALSE                    NA          1741440600
#> 29    FALSE        FALSE                    NA          1742045400
#> 30    FALSE        FALSE                    NA          1743527700
#> 31    FALSE        FALSE                    NA          1743856200
#> 32    FALSE        FALSE                    NA          1744461000
#> 33    FALSE        FALSE                    NA          1745065800
#> 34    FALSE        FALSE                    NA          1745670600
#> 35    FALSE        FALSE                    NA          1746275400
#> 36    FALSE        FALSE                    NA          1746880200
#> 37    FALSE        FALSE                    NA          1747571400
#> 38    FALSE        FALSE                    NA          1748179800
#>    deadline_time_game_offset highest_score is_previous is_current is_next
#> 1                          0           127       FALSE      FALSE   FALSE
#> 2                          0           154       FALSE      FALSE   FALSE
#> 3                          0           130       FALSE      FALSE   FALSE
#> 4                          0           120       FALSE      FALSE   FALSE
#> 5                          0           131       FALSE      FALSE   FALSE
#> 6                          0           142       FALSE      FALSE   FALSE
#> 7                          0           130       FALSE      FALSE   FALSE
#> 8                          0            98       FALSE      FALSE   FALSE
#> 9                          0           117       FALSE      FALSE   FALSE
#> 10                         0           111       FALSE      FALSE   FALSE
#> 11                         0           136       FALSE      FALSE   FALSE
#> 12                         0           128       FALSE      FALSE   FALSE
#> 13                         0           144       FALSE      FALSE   FALSE
#> 14                         0           132       FALSE      FALSE   FALSE
#> 15                         0           119       FALSE      FALSE   FALSE
#> 16                         0           119       FALSE      FALSE   FALSE
#> 17                         0           151       FALSE      FALSE   FALSE
#> 18                         0           118       FALSE      FALSE   FALSE
#> 19                         0           142       FALSE      FALSE   FALSE
#> 20                         0           136       FALSE      FALSE   FALSE
#> 21                         0           148        TRUE      FALSE   FALSE
#> 22                         0           128       FALSE       TRUE   FALSE
#> 23                         0            NA       FALSE      FALSE    TRUE
#> 24                         0            NA       FALSE      FALSE   FALSE
#> 25                         0            NA       FALSE      FALSE   FALSE
#> 26                         0            NA       FALSE      FALSE   FALSE
#> 27                         0            NA       FALSE      FALSE   FALSE
#> 28                         0            NA       FALSE      FALSE   FALSE
#> 29                         0            NA       FALSE      FALSE   FALSE
#> 30                         0            NA       FALSE      FALSE   FALSE
#> 31                         0            NA       FALSE      FALSE   FALSE
#> 32                         0            NA       FALSE      FALSE   FALSE
#> 33                         0            NA       FALSE      FALSE   FALSE
#> 34                         0            NA       FALSE      FALSE   FALSE
#> 35                         0            NA       FALSE      FALSE   FALSE
#> 36                         0            NA       FALSE      FALSE   FALSE
#> 37                         0            NA       FALSE      FALSE   FALSE
#> 38                         0            NA       FALSE      FALSE   FALSE
#>    cup_leagues_created h2h_ko_matches_created can_enter can_manage released
#> 1                FALSE                  FALSE     FALSE      FALSE     TRUE
#> 2                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 3                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 4                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 5                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 6                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 7                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 8                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 9                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 10                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 11                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 12                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 13                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 14                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 15                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 16                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 17                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 18                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 19                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 20                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 21                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 22                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 23                TRUE                   TRUE      TRUE       TRUE     TRUE
#> 24               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 25               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 26               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 27               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 28               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 29               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 30               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 31               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 32               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 33               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 34               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 35               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 36               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 37               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 38               FALSE                  FALSE     FALSE      FALSE     TRUE
#>    ranked_count overrides.element_types overrides.pick_multiplier
#> 1       8597356                    NULL                        NA
#> 2       9543353                    NULL                        NA
#> 3       9912161                    NULL                        NA
#> 4      10178753                    NULL                        NA
#> 5      10279610                    NULL                        NA
#> 6      10369689                    NULL                        NA
#> 7      10466714                    NULL                        NA
#> 8      10552679                    NULL                        NA
#> 9      10594653                    NULL                        NA
#> 10     10648436                    NULL                        NA
#> 11     10693676                    NULL                        NA
#> 12     10751495                    NULL                        NA
#> 13     10786063                    NULL                        NA
#> 14     10821193                    NULL                        NA
#> 15     10854873                    NULL                        NA
#> 16     10906389                    NULL                        NA
#> 17     10946145                    NULL                        NA
#> 18     10984600                    NULL                        NA
#> 19     11011909                    NULL                        NA
#> 20     11058766                    NULL                        NA
#> 21     11114287                    NULL                        NA
#> 22     11136058                    NULL                        NA
#> 23            0                    NULL                        NA
#> 24            0                    NULL                        NA
#> 25            0                    NULL                        NA
#> 26            0                    NULL                        NA
#> 27            0                    NULL                        NA
#> 28            0                    NULL                        NA
#> 29            0                    NULL                        NA
#> 30            0                    NULL                        NA
#> 31            0                    NULL                        NA
#> 32            0                    NULL                        NA
#> 33            0                    NULL                        NA
#> 34            0                    NULL                        NA
#> 35            0                    NULL                        NA
#> 36            0                    NULL                        NA
#> 37            0                    NULL                        NA
#> 38            0                    NULL                        NA
#>                                                        chip_plays most_selected
#> 1                                     bboost, 3xc, 144974, 221430           401
#> 2  bboost, freehit, wildcard, 3xc, 118782, 102788, 217521, 612384           401
#> 3   bboost, freehit, wildcard, 3xc, 87334, 140192, 656652, 186907           351
#> 4   bboost, freehit, wildcard, 3xc, 83637, 125561, 940783, 199688           351
#> 5     bboost, freehit, wildcard, 3xc, 66447, 79498, 311710, 71288           351
#> 6     bboost, freehit, wildcard, 3xc, 62665, 84229, 663489, 83406           351
#> 7    bboost, freehit, wildcard, 3xc, 54778, 81146, 308835, 140205           351
#> 8    bboost, freehit, wildcard, 3xc, 65446, 64301, 221063, 111602           351
#> 9    bboost, freehit, wildcard, 3xc, 51136, 71811, 158486, 330035           351
#> 10    bboost, freehit, wildcard, 3xc, 75823, 70729, 162964, 85359           351
#> 11    bboost, freehit, wildcard, 3xc, 79875, 84689, 138253, 52784           351
#> 12   bboost, freehit, wildcard, 3xc, 59590, 89443, 252772, 111773           351
#> 13    bboost, freehit, wildcard, 3xc, 57138, 71881, 144095, 44722           328
#> 14    bboost, freehit, wildcard, 3xc, 58057, 60749, 141083, 80792           328
#> 15    bboost, freehit, wildcard, 3xc, 46111, 102956, 98439, 72929           182
#> 16   bboost, freehit, wildcard, 3xc, 72296, 56121, 120570, 100883           328
#> 17    bboost, freehit, wildcard, 3xc, 80801, 59644, 136974, 55811           182
#> 18   bboost, freehit, wildcard, 3xc, 97727, 83399, 374001, 202796           328
#> 19    bboost, freehit, wildcard, 3xc, 83841, 62471, 427439, 67924           182
#> 20   bboost, freehit, wildcard, 3xc, 73435, 67748, 412768, 141464           328
#> 21    bboost, freehit, wildcard, 3xc, 89332, 63224, 286969, 51092           328
#> 22   bboost, freehit, wildcard, 3xc, 102686, 35776, 166215, 45688           328
#> 23                                                           NULL            NA
#> 24                                                           NULL            NA
#> 25                                                           NULL            NA
#> 26                                                           NULL            NA
#> 27                                                           NULL            NA
#> 28                                                           NULL            NA
#> 29                                                           NULL            NA
#> 30                                                           NULL            NA
#> 31                                                           NULL            NA
#> 32                                                           NULL            NA
#> 33                                                           NULL            NA
#> 34                                                           NULL            NA
#> 35                                                           NULL            NA
#> 36                                                           NULL            NA
#> 37                                                           NULL            NA
#> 38                                                           NULL            NA
#>    most_transferred_in top_element top_element_info.id top_element_info.points
#> 1                   27         328                 328                      14
#> 2                  594         177                 177                      20
#> 3                  177         328                 328                      17
#> 4                  129         185                 185                      15
#> 5                   58         180                 180                      16
#> 6                  327         182                 182                      25
#> 7                  182         512                 512                      18
#> 8                  491         228                 228                      14
#> 9                   99          99                  99                      15
#> 10                  99          82                  82                      16
#> 11                 447         366                 366                      17
#> 12                 366         541                 541                      16
#> 13                 541         106                 106                      23
#> 14                 180         328                 328                      18
#> 15                  17         182                 182                      18
#> 16                 401         585                 585                      17
#> 17                 364         328                 328                      21
#> 18                 401         403                 403                      13
#> 19                 401         328                 328                      16
#> 20                 447          99                  99                      19
#> 21                 398         364                 364                      20
#> 22                 447          71                  71                      22
#> 23                  NA          NA                  NA                      NA
#> 24                  NA          NA                  NA                      NA
#> 25                  NA          NA                  NA                      NA
#> 26                  NA          NA                  NA                      NA
#> 27                  NA          NA                  NA                      NA
#> 28                  NA          NA                  NA                      NA
#> 29                  NA          NA                  NA                      NA
#> 30                  NA          NA                  NA                      NA
#> 31                  NA          NA                  NA                      NA
#> 32                  NA          NA                  NA                      NA
#> 33                  NA          NA                  NA                      NA
#> 34                  NA          NA                  NA                      NA
#> 35                  NA          NA                  NA                      NA
#> 36                  NA          NA                  NA                      NA
#> 37                  NA          NA                  NA                      NA
#> 38                  NA          NA                  NA                      NA
#>    transfers_made most_captained most_vice_captained
#> 1               0            351                 351
#> 2        13364453            351                 401
#> 3        23723836            351                 328
#> 4        25930392            351                 328
#> 5        13036058            351                 351
#> 6        21036913            351                 328
#> 7        14594461            351                 328
#> 8        11911632            351                 328
#> 9        10882045            351                 182
#> 10       11763382            351                 182
#> 11       11631910            351                 328
#> 12       13315756            351                 182
#> 13       10651205            328                 182
#> 14       10912413            351                 328
#> 15       10409133            328                 182
#> 16       10586275            328                 182
#> 17       10405730            328                 182
#> 18       13895381            328                 182
#> 19       10957444            328                 182
#> 20       13978697            328                 182
#> 21       11543203            328                 182
#> 22        8284410            328                 182
#> 23              0             NA                  NA
#> 24              0             NA                  NA
#> 25              0             NA                  NA
#> 26              0             NA                  NA
#> 27              0             NA                  NA
#> 28              0             NA                  NA
#> 29              0             NA                  NA
#> 30              0             NA                  NA
#> 31              0             NA                  NA
#> 32              0             NA                  NA
#> 33              0             NA                  NA
#> 34              0             NA                  NA
#> 35              0             NA                  NA
#> 36              0             NA                  NA
#> 37              0             NA                  NA
#> 38              0             NA                  NA
#> 
#> $game_settings
#> $game_settings$league_join_private_max
#> [1] 30
#> 
#> $game_settings$league_join_public_max
#> [1] 5
#> 
#> $game_settings$league_max_size_public_classic
#> [1] 20
#> 
#> $game_settings$league_max_size_public_h2h
#> [1] 16
#> 
#> $game_settings$league_max_size_private_h2h
#> [1] 16
#> 
#> $game_settings$league_max_ko_rounds_private_h2h
#> [1] 3
#> 
#> $game_settings$league_prefix_public
#> [1] "League"
#> 
#> $game_settings$league_points_h2h_win
#> [1] 3
#> 
#> $game_settings$league_points_h2h_lose
#> [1] 0
#> 
#> $game_settings$league_points_h2h_draw
#> [1] 1
#> 
#> $game_settings$league_ko_first_instead_of_random
#> [1] FALSE
#> 
#> $game_settings$cup_start_event_id
#> NULL
#> 
#> $game_settings$cup_stop_event_id
#> NULL
#> 
#> $game_settings$cup_qualifying_method
#> NULL
#> 
#> $game_settings$cup_type
#> NULL
#> 
#> $game_settings$featured_entries
#> list()
#> 
#> $game_settings$element_sell_at_purchase_price
#> [1] FALSE
#> 
#> $game_settings$percentile_ranks
#>  [1]  1  5 10 15 20 25 30 35 40 45 50 55 60 65 70 75 80 95 90 95
#> 
#> $game_settings$underdog_differential
#> [1] 5
#> 
#> $game_settings$squad_squadplay
#> [1] 11
#> 
#> $game_settings$squad_squadsize
#> [1] 15
#> 
#> $game_settings$squad_special_min
#> NULL
#> 
#> $game_settings$squad_special_max
#> NULL
#> 
#> $game_settings$squad_team_limit
#> [1] 3
#> 
#> $game_settings$squad_total_spend
#> [1] 1000
#> 
#> $game_settings$ui_currency_multiplier
#> [1] 10
#> 
#> $game_settings$ui_use_special_shirts
#> [1] FALSE
#> 
#> $game_settings$ui_special_shirt_exclusions
#> list()
#> 
#> $game_settings$stats_form_days
#> [1] 30
#> 
#> $game_settings$sys_vice_captain_enabled
#> [1] TRUE
#> 
#> $game_settings$transfers_cap
#> [1] 20
#> 
#> $game_settings$transfers_sell_on_fee
#> [1] 0.5
#> 
#> $game_settings$max_extra_free_transfers
#> [1] 4
#> 
#> $game_settings$league_h2h_tiebreak_stats
#> [1] "+goals_scored"   "-goals_conceded"
#> 
#> $game_settings$timezone
#> [1] "UTC"
#> 
#> 
#> $game_config
#> $game_config$settings
#> $game_config$settings$entry_per_event
#> [1] FALSE
#> 
#> $game_config$settings$timezone
#> [1] "UTC"
#> 
#> 
#> $game_config$rules
#> $game_config$rules$league_join_private_max
#> [1] 30
#> 
#> $game_config$rules$league_join_public_max
#> [1] 5
#> 
#> $game_config$rules$league_max_size_public_classic
#> [1] 20
#> 
#> $game_config$rules$league_max_size_public_h2h
#> [1] 16
#> 
#> $game_config$rules$league_max_size_private_h2h
#> [1] 16
#> 
#> $game_config$rules$league_max_ko_rounds_private_h2h
#> [1] 3
#> 
#> $game_config$rules$league_prefix_public
#> [1] "League"
#> 
#> $game_config$rules$league_points_h2h_win
#> [1] 3
#> 
#> $game_config$rules$league_points_h2h_lose
#> [1] 0
#> 
#> $game_config$rules$league_points_h2h_draw
#> [1] 1
#> 
#> $game_config$rules$league_ko_first_instead_of_random
#> [1] FALSE
#> 
#> $game_config$rules$cup_start_event_id
#> NULL
#> 
#> $game_config$rules$cup_stop_event_id
#> NULL
#> 
#> $game_config$rules$cup_qualifying_method
#> NULL
#> 
#> $game_config$rules$cup_type
#> NULL
#> 
#> $game_config$rules$featured_entries
#> list()
#> 
#> $game_config$rules$element_sell_at_purchase_price
#> [1] FALSE
#> 
#> $game_config$rules$percentile_ranks
#>  [1]  1  5 10 15 20 25 30 35 40 45 50 55 60 65 70 75 80 95 90 95
#> 
#> $game_config$rules$underdog_differential
#> [1] 5
#> 
#> $game_config$rules$squad_squadplay
#> [1] 11
#> 
#> $game_config$rules$squad_squadsize
#> [1] 15
#> 
#> $game_config$rules$squad_special_min
#> NULL
#> 
#> $game_config$rules$squad_special_max
#> NULL
#> 
#> $game_config$rules$squad_team_limit
#> [1] 3
#> 
#> $game_config$rules$squad_total_spend
#> [1] 1000
#> 
#> $game_config$rules$ui_currency_multiplier
#> [1] 10
#> 
#> $game_config$rules$ui_use_special_shirts
#> [1] FALSE
#> 
#> $game_config$rules$ui_special_shirt_exclusions
#> list()
#> 
#> $game_config$rules$stats_form_days
#> [1] 30
#> 
#> $game_config$rules$sys_vice_captain_enabled
#> [1] TRUE
#> 
#> $game_config$rules$transfers_cap
#> [1] 20
#> 
#> $game_config$rules$transfers_sell_on_fee
#> [1] 0.5
#> 
#> $game_config$rules$max_extra_free_transfers
#> [1] 4
#> 
#> $game_config$rules$league_h2h_tiebreak_stats
#> [1] "+goals_scored"   "-goals_conceded"
#> 
#> 
#> $game_config$scoring
#> $game_config$scoring$long_play
#> [1] 2
#> 
#> $game_config$scoring$short_play
#> [1] 1
#> 
#> $game_config$scoring$goals_conceded
#> $game_config$scoring$goals_conceded$DEF
#> [1] -1
#> 
#> $game_config$scoring$goals_conceded$FWD
#> [1] 0
#> 
#> $game_config$scoring$goals_conceded$GKP
#> [1] -1
#> 
#> $game_config$scoring$goals_conceded$MID
#> [1] 0
#> 
#> $game_config$scoring$goals_conceded$MNG
#> [1] 0
#> 
#> 
#> $game_config$scoring$saves
#> [1] 1
#> 
#> $game_config$scoring$goals_scored
#> $game_config$scoring$goals_scored$DEF
#> [1] 6
#> 
#> $game_config$scoring$goals_scored$FWD
#> [1] 4
#> 
#> $game_config$scoring$goals_scored$GKP
#> [1] 10
#> 
#> $game_config$scoring$goals_scored$MID
#> [1] 5
#> 
#> $game_config$scoring$goals_scored$MNG
#> [1] 0
#> 
#> 
#> $game_config$scoring$assists
#> [1] 3
#> 
#> $game_config$scoring$clean_sheets
#> $game_config$scoring$clean_sheets$DEF
#> [1] 4
#> 
#> $game_config$scoring$clean_sheets$FWD
#> [1] 0
#> 
#> $game_config$scoring$clean_sheets$GKP
#> [1] 4
#> 
#> $game_config$scoring$clean_sheets$MID
#> [1] 1
#> 
#> $game_config$scoring$clean_sheets$MNG
#> [1] 0
#> 
#> 
#> $game_config$scoring$penalties_saved
#> [1] 5
#> 
#> $game_config$scoring$penalties_missed
#> [1] -2
#> 
#> $game_config$scoring$yellow_cards
#> [1] -1
#> 
#> $game_config$scoring$red_cards
#> [1] -3
#> 
#> $game_config$scoring$own_goals
#> [1] -2
#> 
#> $game_config$scoring$bonus
#> [1] 1
#> 
#> $game_config$scoring$bps
#> [1] 0
#> 
#> $game_config$scoring$influence
#> [1] 0
#> 
#> $game_config$scoring$creativity
#> [1] 0
#> 
#> $game_config$scoring$threat
#> [1] 0
#> 
#> $game_config$scoring$ict_index
#> [1] 0
#> 
#> $game_config$scoring$special_multiplier
#> [1] 1
#> 
#> $game_config$scoring$mng_goals_scored
#> $game_config$scoring$mng_goals_scored$DEF
#> [1] 0
#> 
#> $game_config$scoring$mng_goals_scored$FWD
#> [1] 0
#> 
#> $game_config$scoring$mng_goals_scored$GKP
#> [1] 0
#> 
#> $game_config$scoring$mng_goals_scored$MID
#> [1] 0
#> 
#> $game_config$scoring$mng_goals_scored$MNG
#> [1] 1
#> 
#> 
#> $game_config$scoring$mng_clean_sheets
#> $game_config$scoring$mng_clean_sheets$DEF
#> [1] 0
#> 
#> $game_config$scoring$mng_clean_sheets$FWD
#> [1] 0
#> 
#> $game_config$scoring$mng_clean_sheets$GKP
#> [1] 0
#> 
#> $game_config$scoring$mng_clean_sheets$MID
#> [1] 0
#> 
#> $game_config$scoring$mng_clean_sheets$MNG
#> [1] 2
#> 
#> 
#> $game_config$scoring$mng_win
#> $game_config$scoring$mng_win$DEF
#> [1] 0
#> 
#> $game_config$scoring$mng_win$FWD
#> [1] 0
#> 
#> $game_config$scoring$mng_win$GKP
#> [1] 0
#> 
#> $game_config$scoring$mng_win$MID
#> [1] 0
#> 
#> $game_config$scoring$mng_win$MNG
#> [1] 6
#> 
#> 
#> $game_config$scoring$mng_draw
#> $game_config$scoring$mng_draw$DEF
#> [1] 0
#> 
#> $game_config$scoring$mng_draw$FWD
#> [1] 0
#> 
#> $game_config$scoring$mng_draw$GKP
#> [1] 0
#> 
#> $game_config$scoring$mng_draw$MID
#> [1] 0
#> 
#> $game_config$scoring$mng_draw$MNG
#> [1] 3
#> 
#> 
#> $game_config$scoring$mng_loss
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_win
#> $game_config$scoring$mng_underdog_win$DEF
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_win$FWD
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_win$GKP
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_win$MID
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_win$MNG
#> [1] 10
#> 
#> 
#> $game_config$scoring$mng_underdog_draw
#> $game_config$scoring$mng_underdog_draw$DEF
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_draw$FWD
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_draw$GKP
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_draw$MID
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_draw$MNG
#> [1] 5
#> 
#> 
#> 
#> 
#> $phases
#>    id      name start_event stop_event highest_score
#> 1   1   Overall           1         38          1605
#> 2   2    August           1          3           309
#> 3   3 September           4          6           301
#> 4   4   October           7          9           267
#> 5   5  November          10         13           362
#> 6   6  December          14         19           551
#> 7   7   January          20         23           329
#> 8   8  February          24         27            NA
#> 9   9     March          28         29            NA
#> 10 10     April          30         34            NA
#> 11 11       May          35         38            NA
#> 
#> $teams
#>    code draw form id loss           name played points position short_name
#> 1     3    0   NA  1    0        Arsenal      0      0        2        ARS
#> 2     7    0   NA  2    0    Aston Villa      0      0        8        AVL
#> 3    91    0   NA  3    0    Bournemouth      0      0        7        BOU
#> 4    94    0   NA  4    0      Brentford      0      0       11        BRE
#> 5    36    0   NA  5    0       Brighton      0      0        9        BHA
#> 6     8    0   NA  6    0        Chelsea      0      0        4        CHE
#> 7    31    0   NA  7    0 Crystal Palace      0      0       12        CRY
#> 8    11    0   NA  8    0        Everton      0      0       16        EVE
#> 9    54    0   NA  9    0         Fulham      0      0       10        FUL
#> 10   40    0   NA 10    0        Ipswich      0      0       18        IPS
#> 11   13    0   NA 11    0      Leicester      0      0       19        LEI
#> 12   14    0   NA 12    0      Liverpool      0      0        1        LIV
#> 13   43    0   NA 13    0       Man City      0      0        5        MCI
#> 14    1    0   NA 14    0        Man Utd      0      0       13        MUN
#> 15    4    0   NA 15    0      Newcastle      0      0        6        NEW
#> 16   17    0   NA 16    0  Nott'm Forest      0      0        3        NFO
#> 17   20    0   NA 17    0    Southampton      0      0       20        SOU
#> 18    6    0   NA 18    0          Spurs      0      0       15        TOT
#> 19   21    0   NA 19    0       West Ham      0      0       14        WHU
#> 20   39    0   NA 20    0         Wolves      0      0       17        WOL
#>    strength team_division unavailable win strength_overall_home
#> 1         4            NA       FALSE   0                  1250
#> 2         3            NA       FALSE   0                  1110
#> 3         3            NA       FALSE   0                  1140
#> 4         3            NA       FALSE   0                  1080
#> 5         3            NA       FALSE   0                  1110
#> 6         4            NA       FALSE   0                  1255
#> 7         3            NA       FALSE   0                  1120
#> 8         3            NA       FALSE   0                  1095
#> 9         3            NA       FALSE   0                  1125
#> 10        3            NA       FALSE   0                  1075
#> 11        3            NA       FALSE   0                  1070
#> 12        5            NA       FALSE   0                  1350
#> 13        4            NA       FALSE   0                  1220
#> 14        3            NA       FALSE   0                  1125
#> 15        3            NA       FALSE   0                  1115
#> 16        3            NA       FALSE   0                  1130
#> 17        2            NA       FALSE   0                  1055
#> 18        4            NA       FALSE   0                  1220
#> 19        3            NA       FALSE   0                  1100
#> 20        3            NA       FALSE   0                  1085
#>    strength_overall_away strength_attack_home strength_attack_away
#> 1                   1360                 1260                 1370
#> 2                   1240                 1080                 1160
#> 3                   1175                 1080                 1140
#> 4                   1225                 1070                 1100
#> 5                   1150                 1080                 1100
#> 6                   1255                 1150                 1150
#> 7                   1125                 1130                 1140
#> 8                   1125                 1130                 1150
#> 9                   1140                 1140                 1140
#> 10                  1075                 1070                 1070
#> 11                  1090                 1050                 1080
#> 12                  1360                 1340                 1360
#> 13                  1290                 1140                 1230
#> 14                  1155                 1140                 1140
#> 15                  1220                 1100                 1150
#> 16                  1175                 1140                 1230
#> 17                  1075                 1060                 1060
#> 18                  1245                 1140                 1140
#> 19                  1090                 1080                 1070
#> 20                  1095                 1070                 1070
#>    strength_defence_home strength_defence_away pulse_id
#> 1                   1240                  1350        1
#> 2                   1140                  1320        2
#> 3                   1190                  1210      127
#> 4                   1090                  1350      130
#> 5                   1140                  1200      131
#> 6                   1360                  1360        4
#> 7                   1110                  1110        6
#> 8                   1060                  1100        7
#> 9                   1110                  1160       34
#> 10                  1080                  1080        8
#> 11                  1090                  1110       26
#> 12                  1360                  1360       10
#> 13                  1300                  1350       11
#> 14                  1110                  1170       12
#> 15                  1130                  1290       23
#> 16                  1130                  1130       15
#> 17                  1050                  1090       20
#> 18                  1300                  1350       21
#> 19                  1100                  1130       25
#> 20                  1100                  1120       38
#>