R6 Class Representing a User

The [FPL$get_user()] method returns an object of class `User`.

Public fields

user

A `list` of user details.

history

A `list` of `data.frame`s containing the historic gameweek performance data for the user.

cup

A `list` of data related to the user's cup performance.

picks

A `list` of the user's picks information for each gameweek.

transfers

A `list` of the user's weekly transfers.

Methods

Public methods

User$new()

User$get_gameweek_history()

User$get_season_history()

User$get_chips_history()

User$get_picks()

User$get_cup_status()

User$get_cup_matches()

User$get_active_chips()

User$get_automatic_substitutions()

User$get_user_history()

User$get_transfers()

User$print()

User$clone()

Method new()

Instantiate the `User` class.

Usage

User$new(user)

Arguments

user

`list`. Data for a single FPL user.

Returns

An R6 `User` object.

Examples

fpl <- FPL$new()
user <- fpl$get_user(555690)

Method get_gameweek_history()

Get the gameweek history for the user.

Usage

User$get_gameweek_history(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A `data.frame`.

Examples

user$get_gameweek_history()

Method get_season_history()

Get the seasonal history for the user.

Usage

User$get_season_history()

Returns

A `data.frame`.

Examples

user$get_season_history()

Method get_chips_history()

Get the chip history for the user.

Usage

User$get_chips_history(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A `data.frame`.

Examples

user$get_gameweek_history()

Method get_picks()

Get the user's picks for the chosen `gameweek`(s).

Usage

User$get_picks(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek picks to return (default: `NULL` returns all gameweeks).

Returns

A `list` containing the user's picks details for the chosen gameweek(s).

Examples

user$get_picks()

Method get_cup_status()

Get the user's cup status.

Usage

User$get_cup_status()

Returns

A `list` of cup status information.

Examples

user$get_cup_status()

Method get_cup_matches()

Get the user's cup matches.

Usage

User$get_cup_matches(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A `list` of cup match information.

Examples

user$get_cup_matches()

Method get_active_chips()

Get the user's active chip for each gameweek or the active chip of the chosen `gameweek`(s).

Usage

User$get_active_chips(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A vector of active chips.

Examples

user$get_active_chips()

Method get_automatic_substitutions()

Get a `list` of automatic substitutions for the chosen `gameweek`(s).

Usage

User$get_automatic_substitutions(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A vector of active chips.

Examples

user$get_automatic_substitutions()

Method get_user_history()

Get a `list` of the user's `gameweek` history.

Usage

User$get_user_history(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A `list` of `data.frame`s of the user's gameweek history.

Examples

user$get_user_history()

Method get_transfers()

Get a `list` of all the user's transfers, or a `list` of transfers made in the given `gameweek`(s).

Usage

User$get_transfers(gameweek = NULL)

Arguments

gameweek

`numeric(n)`. The gameweek history to return (default: `NULL` returns all gameweeks).

Returns

A `list` of the user's transfers.

Examples

user$get_transfers()

Method print()

Pretty printing of the `User`.

Usage

User$print(...)

Arguments

...

Not used.

Method clone()

The objects of this class are cloneable with this method.

Usage

User$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `User$new`
## ------------------------------------------------

fpl <- FPL$new()
user <- fpl$get_user(555690)

## ------------------------------------------------
## Method `User$get_gameweek_history`
## ------------------------------------------------

user$get_gameweek_history()

#>   event points total_points   rank rank_sort overall_rank bank value
#> 1     1     61           61 738822    747511       738822   30  1000
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               0

## ------------------------------------------------
## Method `User$get_season_history`
## ------------------------------------------------

user$get_season_history()

#>   season_name total_points   rank
#> 1     2018/19         2198 278647
#> 2     2019/20         2205 273955

## ------------------------------------------------
## Method `User$get_chips_history`
## ------------------------------------------------

user$get_gameweek_history()

#>   event points total_points   rank rank_sort overall_rank bank value
#> 1     1     61           61 738822    747511       738822   30  1000
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               0

## ------------------------------------------------
## Method `User$get_picks`
## ------------------------------------------------

user$get_picks()

#> [[1]]
#> [[1]]$active_chip
#> NULL
#> 
#> [[1]]$automatic_subs
#> list()
#> 
#> [[1]]$entry_history
#>   event points total_points   rank rank_sort overall_rank bank value
#> 1     1     61           61 738822    747511       738822   30  1000
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               0
#> 
#> [[1]]$picks
#>    element position multiplier is_captain is_vice_captain
#> 1      363        1          1      FALSE           FALSE
#> 2      259        2          1      FALSE           FALSE
#> 3      467        3          1      FALSE           FALSE
#> 4      457        4          1      FALSE           FALSE
#> 5      254        5          2       TRUE           FALSE
#> 6      235        6          1      FALSE           FALSE
#> 7      338        7          1      FALSE           FALSE
#> 8        4        8          1      FALSE            TRUE
#> 9      184        9          1      FALSE           FALSE
#> 10     377       10          1      FALSE           FALSE
#> 11     117       11          1      FALSE           FALSE
#> 12      35       12          0      FALSE           FALSE
#> 13     196       13          0      FALSE           FALSE
#> 14      76       14          0      FALSE           FALSE
#> 15      97       15          0      FALSE           FALSE
#> 
#> 

## ------------------------------------------------
## Method `User$get_cup_status`
## ------------------------------------------------

user$get_cup_status()

#> $qualification_event
#> [1] 16
#> 
#> $qualification_numbers
#> [1] 4194304
#> 
#> $qualification_rank
#> NULL
#> 
#> $qualification_state
#> [1] "NOT_STARTED"
#> 

## ------------------------------------------------
## Method `User$get_cup_matches`
## ------------------------------------------------

user$get_cup_matches()

#> list()

## ------------------------------------------------
## Method `User$get_active_chips`
## ------------------------------------------------

user$get_active_chips()

#> [[1]]
#> NULL
#> 

## ------------------------------------------------
## Method `User$get_automatic_substitutions`
## ------------------------------------------------

user$get_automatic_substitutions()

#> [[1]]
#> list()
#> 

## ------------------------------------------------
## Method `User$get_user_history`
## ------------------------------------------------

user$get_user_history()

#> [[1]]
#>   event points total_points   rank rank_sort overall_rank bank value
#> 1     1     61           61 738822    747511       738822   30  1000
#>   event_transfers event_transfers_cost points_on_bench
#> 1               0                    0               0
#> 

## ------------------------------------------------
## Method `User$get_transfers`
## ------------------------------------------------

user$get_transfers()

#> list()