Get the current the stats from the fantasy API

Get the current the stats from the fantasy API

Usage

get_premier_stats()

Value

data frame with all players and variables

Examples

# Get the stats from online
weekly_stats <- get_premier_stats()
head(weekly_stats) # print stats
#> $chips
#>   id     name number start_event stop_event chip_type overrides.squad_squadsize
#> 1  1 wildcard      1           2         19  transfer                        NA
#> 2  2 wildcard      1          20         38  transfer                        NA
#> 3  3  freehit      1           2         38  transfer                        NA
#> 4  4   bboost      1           1         38      team                        NA
#> 5  5      3xc      1           1         38      team                        NA
#> 6  6  manager      1          24         38  transfer                        16
#>                                         overrides.element_types
#> 1                                                          NULL
#> 2                                                          NULL
#> 3                                                          NULL
#> 4                                                          NULL
#> 5                                                          NULL
#> 6 5, Managers, MNG, Manager, MNG, 1, NA, NA, 0, 0, TRUE, 16, 20
#>   overrides.pick_multiplier
#> 1                        NA
#> 2                        NA
#> 3                        NA
#> 4                        NA
#> 5                        NA
#> 6                        NA
#> 
#> $events
#>    id        name        deadline_time release_time average_entry_score
#> 1   1  Gameweek 1 2024-08-16T17:30:00Z           NA                  57
#> 2   2  Gameweek 2 2024-08-24T10:00:00Z           NA                  69
#> 3   3  Gameweek 3 2024-08-31T10:00:00Z           NA                  64
#> 4   4  Gameweek 4 2024-09-14T10:00:00Z           NA                  51
#> 5   5  Gameweek 5 2024-09-21T10:00:00Z           NA                  58
#> 6   6  Gameweek 6 2024-09-28T10:00:00Z           NA                  50
#> 7   7  Gameweek 7 2024-10-05T10:00:00Z           NA                  46
#> 8   8  Gameweek 8 2024-10-19T10:00:00Z           NA                  36
#> 9   9  Gameweek 9 2024-10-25T17:30:00Z           NA                  54
#> 10 10 Gameweek 10 2024-11-02T11:00:00Z           NA                  39
#> 11 11 Gameweek 11 2024-11-09T13:30:00Z           NA                  49
#> 12 12 Gameweek 12 2024-11-23T11:00:00Z           NA                  49
#> 13 13 Gameweek 13 2024-11-29T18:30:00Z           NA                  60
#> 14 14 Gameweek 14 2024-12-03T18:00:00Z           NA                  58
#> 15 15 Gameweek 15 2024-12-07T11:00:00Z           NA                  50
#> 16 16 Gameweek 16 2024-12-14T13:30:00Z           NA                  47
#> 17 17 Gameweek 17 2024-12-21T11:00:00Z           NA                  60
#> 18 18 Gameweek 18 2024-12-26T11:00:00Z           NA                  51
#> 19 19 Gameweek 19 2024-12-29T13:00:00Z           NA                  66
#> 20 20 Gameweek 20 2025-01-04T11:00:00Z           NA                  60
#> 21 21 Gameweek 21 2025-01-14T18:00:00Z           NA                  55
#> 22 22 Gameweek 22 2025-01-18T11:00:00Z           NA                  46
#> 23 23 Gameweek 23 2025-01-25T13:30:00Z           NA                  59
#> 24 24 Gameweek 24 2025-02-01T11:00:00Z           NA                  87
#> 25 25 Gameweek 25 2025-02-14T18:30:00Z           NA                  74
#> 26 26 Gameweek 26 2025-02-21T18:30:00Z           NA                  62
#> 27 27 Gameweek 27 2025-02-25T18:00:00Z           NA                  53
#> 28 28 Gameweek 28 2025-03-08T11:00:00Z           NA                  52
#> 29 29 Gameweek 29 2025-03-15T13:30:00Z           NA                  40
#> 30 30 Gameweek 30 2025-04-01T17:15:00Z           NA                  44
#> 31 31 Gameweek 31 2025-04-05T10:00:00Z           NA                  34
#> 32 32 Gameweek 32 2025-04-12T10:00:00Z           NA                  56
#> 33 33 Gameweek 33 2025-04-19T12:30:00Z           NA                   0
#> 34 34 Gameweek 34 2025-04-26T10:00:00Z           NA                   0
#> 35 35 Gameweek 35 2025-05-02T17:30:00Z           NA                   0
#> 36 36 Gameweek 36 2025-05-10T10:00:00Z           NA                   0
#> 37 37 Gameweek 37 2025-05-18T12:30:00Z           NA                   0
#> 38 38 Gameweek 38 2025-05-25T13:30:00Z           NA                   0
#>    finished data_checked highest_scoring_entry deadline_time_epoch
#> 1      TRUE         TRUE               3546234          1723829400
#> 2      TRUE         TRUE               9442126          1724493600
#> 3      TRUE         TRUE                430195          1725098400
#> 4      TRUE         TRUE               3560750          1726308000
#> 5      TRUE         TRUE               1773336          1726912800
#> 6      TRUE         TRUE               7479996          1727517600
#> 7      TRUE         TRUE                759474          1728122400
#> 8      TRUE         TRUE               5482232          1729332000
#> 9      TRUE         TRUE               2039494          1729877400
#> 10     TRUE         TRUE               1689659          1730545200
#> 11     TRUE         TRUE              10668888          1731159000
#> 12     TRUE         TRUE               7440806          1732359600
#> 13     TRUE         TRUE               1929511          1732905000
#> 14     TRUE         TRUE               2964307          1733248800
#> 15     TRUE         TRUE               7399737          1733569200
#> 16     TRUE         TRUE               5142221          1734183000
#> 17     TRUE         TRUE               4917384          1734778800
#> 18     TRUE         TRUE               2701866          1735210800
#> 19     TRUE         TRUE               3191421          1735477200
#> 20     TRUE         TRUE               1247541          1735988400
#> 21     TRUE         TRUE                434390          1736877600
#> 22     TRUE         TRUE               2700271          1737198000
#> 23     TRUE         TRUE              11007664          1737811800
#> 24     TRUE         TRUE               5428593          1738407600
#> 25     TRUE         TRUE               2065637          1739557800
#> 26     TRUE         TRUE               1546737          1740162600
#> 27     TRUE         TRUE              10980207          1740506400
#> 28     TRUE         TRUE               4865928          1741431600
#> 29     TRUE         TRUE                885345          1742045400
#> 30     TRUE         TRUE              11409817          1743527700
#> 31     TRUE         TRUE               2430340          1743847200
#> 32     TRUE         TRUE                994851          1744452000
#> 33    FALSE        FALSE                    NA          1745065800
#> 34    FALSE        FALSE                    NA          1745661600
#> 35    FALSE        FALSE                    NA          1746207000
#> 36    FALSE        FALSE                    NA          1746871200
#> 37    FALSE        FALSE                    NA          1747571400
#> 38    FALSE        FALSE                    NA          1748179800
#>    deadline_time_game_offset highest_score is_previous is_current is_next
#> 1                          0           127       FALSE      FALSE   FALSE
#> 2                          0           154       FALSE      FALSE   FALSE
#> 3                          0           130       FALSE      FALSE   FALSE
#> 4                          0           120       FALSE      FALSE   FALSE
#> 5                          0           131       FALSE      FALSE   FALSE
#> 6                          0           142       FALSE      FALSE   FALSE
#> 7                          0           130       FALSE      FALSE   FALSE
#> 8                          0            98       FALSE      FALSE   FALSE
#> 9                          0           117       FALSE      FALSE   FALSE
#> 10                         0           111       FALSE      FALSE   FALSE
#> 11                         0           136       FALSE      FALSE   FALSE
#> 12                         0           128       FALSE      FALSE   FALSE
#> 13                         0           144       FALSE      FALSE   FALSE
#> 14                         0           132       FALSE      FALSE   FALSE
#> 15                         0           119       FALSE      FALSE   FALSE
#> 16                         0           119       FALSE      FALSE   FALSE
#> 17                         0           151       FALSE      FALSE   FALSE
#> 18                         0           118       FALSE      FALSE   FALSE
#> 19                         0           142       FALSE      FALSE   FALSE
#> 20                         0           136       FALSE      FALSE   FALSE
#> 21                         0           148       FALSE      FALSE   FALSE
#> 22                         0           128       FALSE      FALSE   FALSE
#> 23                         0           129       FALSE      FALSE   FALSE
#> 24                         0           201       FALSE      FALSE   FALSE
#> 25                         0           155       FALSE      FALSE   FALSE
#> 26                         0           140       FALSE      FALSE   FALSE
#> 27                         0           120       FALSE      FALSE   FALSE
#> 28                         0           122       FALSE      FALSE   FALSE
#> 29                         0           129       FALSE      FALSE   FALSE
#> 30                         0           108       FALSE      FALSE   FALSE
#> 31                         0           109        TRUE      FALSE   FALSE
#> 32                         0           168       FALSE       TRUE   FALSE
#> 33                         0            NA       FALSE      FALSE    TRUE
#> 34                         0            NA       FALSE      FALSE   FALSE
#> 35                         0            NA       FALSE      FALSE   FALSE
#> 36                         0            NA       FALSE      FALSE   FALSE
#> 37                         0            NA       FALSE      FALSE   FALSE
#> 38                         0            NA       FALSE      FALSE   FALSE
#>    cup_leagues_created h2h_ko_matches_created can_enter can_manage released
#> 1                FALSE                  FALSE     FALSE      FALSE     TRUE
#> 2                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 3                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 4                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 5                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 6                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 7                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 8                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 9                 TRUE                   TRUE     FALSE      FALSE     TRUE
#> 10                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 11                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 12                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 13                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 14                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 15                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 16                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 17                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 18                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 19                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 20                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 21                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 22                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 23                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 24                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 25                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 26                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 27                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 28                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 29                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 30                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 31                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 32                TRUE                   TRUE     FALSE      FALSE     TRUE
#> 33                TRUE                   TRUE      TRUE       TRUE     TRUE
#> 34               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 35               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 36               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 37               FALSE                  FALSE     FALSE      FALSE     TRUE
#> 38               FALSE                  FALSE     FALSE      FALSE     TRUE
#>    ranked_count overrides.element_types overrides.pick_multiplier
#> 1       8597356                    NULL                        NA
#> 2       9543353                    NULL                        NA
#> 3       9912161                    NULL                        NA
#> 4      10178753                    NULL                        NA
#> 5      10279610                    NULL                        NA
#> 6      10369689                    NULL                        NA
#> 7      10466714                    NULL                        NA
#> 8      10552679                    NULL                        NA
#> 9      10594653                    NULL                        NA
#> 10     10648436                    NULL                        NA
#> 11     10693676                    NULL                        NA
#> 12     10751495                    NULL                        NA
#> 13     10786063                    NULL                        NA
#> 14     10821193                    NULL                        NA
#> 15     10854873                    NULL                        NA
#> 16     10906389                    NULL                        NA
#> 17     10946145                    NULL                        NA
#> 18     10984600                    NULL                        NA
#> 19     11011909                    NULL                        NA
#> 20     11058766                    NULL                        NA
#> 21     11114287                    NULL                        NA
#> 22     11136058                    NULL                        NA
#> 23     11178740                    NULL                        NA
#> 24     11216322                    NULL                        NA
#> 25     11262176                    NULL                        NA
#> 26     11285417                    NULL                        NA
#> 27     11301288                    NULL                        NA
#> 28     11328678                    NULL                        NA
#> 29     11344621                    NULL                        NA
#> 30     11367765                    NULL                        NA
#> 31     11373405                    NULL                        NA
#> 32     11382704                    NULL                        NA
#> 33            0                    NULL                        NA
#> 34            0                    NULL                        NA
#> 35            0                    NULL                        NA
#> 36            0                    NULL                        NA
#> 37            0                    NULL                        NA
#> 38            0                    NULL                        NA
#>                                                                           chip_plays
#> 1                                                        bboost, 3xc, 144974, 221430
#> 2                     bboost, freehit, wildcard, 3xc, 118782, 102788, 217521, 612384
#> 3                      bboost, freehit, wildcard, 3xc, 87334, 140192, 656652, 186907
#> 4                      bboost, freehit, wildcard, 3xc, 83637, 125561, 940783, 199688
#> 5                        bboost, freehit, wildcard, 3xc, 66447, 79498, 311710, 71288
#> 6                        bboost, freehit, wildcard, 3xc, 62665, 84229, 663489, 83406
#> 7                       bboost, freehit, wildcard, 3xc, 54778, 81146, 308835, 140205
#> 8                       bboost, freehit, wildcard, 3xc, 65446, 64301, 221063, 111602
#> 9                       bboost, freehit, wildcard, 3xc, 51136, 71811, 158486, 330035
#> 10                       bboost, freehit, wildcard, 3xc, 75823, 70729, 162964, 85359
#> 11                       bboost, freehit, wildcard, 3xc, 79875, 84689, 138253, 52784
#> 12                      bboost, freehit, wildcard, 3xc, 59590, 89443, 252772, 111773
#> 13                       bboost, freehit, wildcard, 3xc, 57138, 71881, 144095, 44722
#> 14                       bboost, freehit, wildcard, 3xc, 58057, 60749, 141083, 80792
#> 15                       bboost, freehit, wildcard, 3xc, 46111, 102956, 98439, 72929
#> 16                      bboost, freehit, wildcard, 3xc, 72296, 56121, 120570, 100883
#> 17                       bboost, freehit, wildcard, 3xc, 80801, 59644, 136974, 55811
#> 18                      bboost, freehit, wildcard, 3xc, 97727, 83399, 374001, 202796
#> 19                       bboost, freehit, wildcard, 3xc, 83841, 62471, 427439, 67924
#> 20                      bboost, freehit, wildcard, 3xc, 73435, 67748, 412768, 141464
#> 21                       bboost, freehit, wildcard, 3xc, 89332, 63224, 286969, 51092
#> 22                      bboost, freehit, wildcard, 3xc, 102686, 35776, 166215, 45688
#> 23                     bboost, freehit, wildcard, 3xc, 125318, 41076, 202189, 145561
#> 24 bboost, freehit, wildcard, 3xc, manager, 111461, 123991, 129674, 1139007, 1177726
#> 25     bboost, freehit, wildcard, 3xc, manager, 82673, 92674, 127582, 636104, 901391
#> 26      bboost, freehit, wildcard, 3xc, manager, 54221, 28439, 103756, 18604, 390248
#> 27      bboost, freehit, wildcard, 3xc, manager, 48902, 29345, 104007, 23870, 352459
#> 28     bboost, freehit, wildcard, 3xc, manager, 71044, 35554, 115172, 154573, 293173
#> 29    bboost, freehit, wildcard, 3xc, manager, 14258, 1189224, 221540, 14819, 105465
#> 30      bboost, freehit, wildcard, 3xc, manager, 70180, 33497, 658984, 32218, 276833
#> 31      bboost, freehit, wildcard, 3xc, manager, 30471, 35898, 476122, 14043, 275454
#> 32    bboost, freehit, wildcard, 3xc, manager, 208952, 77639, 279508, 447125, 286588
#> 33                        bboost, freehit, wildcard, 3xc, 73410, 23855, 15302, 18684
#> 34                                                                              NULL
#> 35                                                                              NULL
#> 36                                                                              NULL
#> 37                                                                              NULL
#> 38                                                                              NULL
#>    most_selected most_transferred_in top_element top_element_info.id
#> 1            401                  27         328                 328
#> 2            401                 594         177                 177
#> 3            351                 177         328                 328
#> 4            351                 129         185                 185
#> 5            351                  58         180                 180
#> 6            351                 327         182                 182
#> 7            351                 182         512                 512
#> 8            351                 491         228                 228
#> 9            351                  99          99                  99
#> 10           351                  99          82                  82
#> 11           351                 447         366                 366
#> 12           351                 366         541                 541
#> 13           328                 541         106                 106
#> 14           328                 180         328                 328
#> 15           182                  17         182                 182
#> 16           328                 401         585                 585
#> 17           182                 364         328                 328
#> 18           328                 401         403                 403
#> 19           182                 401         328                 328
#> 20           328                 447          99                  99
#> 21           328                 398         364                 364
#> 22           328                 447          71                  71
#> 23           328                  71          74                  74
#> 24           328                 321         328                 328
#> 25           328                  71         328                 328
#> 26           328                  71         762                 762
#> 27           328                 541         163                 163
#> 28           328                 211         163                 163
#> 29           328                 366         366                 366
#> 30           328                 755         762                 762
#> 31           328                  17         491                 491
#> 32           328                 402         392                 392
#> 33            NA                  NA          NA                  NA
#> 34            NA                  NA          NA                  NA
#> 35            NA                  NA          NA                  NA
#> 36            NA                  NA          NA                  NA
#> 37            NA                  NA          NA                  NA
#> 38            NA                  NA          NA                  NA
#>    top_element_info.points transfers_made most_captained most_vice_captained
#> 1                       14              0            351                 351
#> 2                       20       13364453            351                 401
#> 3                       17       23723836            351                 328
#> 4                       15       25930392            351                 328
#> 5                       16       13036058            351                 351
#> 6                       25       21036913            351                 328
#> 7                       18       14594461            351                 328
#> 8                       14       11911632            351                 328
#> 9                       15       10882045            351                 182
#> 10                      16       11763382            351                 182
#> 11                      17       11631910            351                 328
#> 12                      16       13315756            351                 182
#> 13                      23       10651205            328                 182
#> 14                      18       10912413            351                 328
#> 15                      18       10409133            328                 182
#> 16                      17       10586275            328                 182
#> 17                      21       10405730            328                 182
#> 18                      13       13895381            328                 182
#> 19                      16       10957444            328                 182
#> 20                      19       13978697            328                 182
#> 21                      20       11543203            328                 182
#> 22                      22        8284410            328                 182
#> 23                      21       10475856            328                 401
#> 24                      29       11002861            328                 401
#> 25                      20       11066698            328                 182
#> 26                      15        8907059            328                 182
#> 27                      15        7849779            328                 182
#> 28                      15        8911975            328                 182
#> 29                      17       21126461            328                 182
#> 30                      12       15567710            328                 401
#> 31                      18       11309600            328                 401
#> 32                      27       10134301            328                 328
#> 33                      NA         973059             NA                  NA
#> 34                      NA              0             NA                  NA
#> 35                      NA              0             NA                  NA
#> 36                      NA              0             NA                  NA
#> 37                      NA              0             NA                  NA
#> 38                      NA              0             NA                  NA
#> 
#> $game_settings
#> $game_settings$league_join_private_max
#> [1] 30
#> 
#> $game_settings$league_join_public_max
#> [1] 5
#> 
#> $game_settings$league_max_size_public_classic
#> [1] 20
#> 
#> $game_settings$league_max_size_public_h2h
#> [1] 16
#> 
#> $game_settings$league_max_size_private_h2h
#> [1] 16
#> 
#> $game_settings$league_max_ko_rounds_private_h2h
#> [1] 3
#> 
#> $game_settings$league_prefix_public
#> [1] "League"
#> 
#> $game_settings$league_points_h2h_win
#> [1] 3
#> 
#> $game_settings$league_points_h2h_lose
#> [1] 0
#> 
#> $game_settings$league_points_h2h_draw
#> [1] 1
#> 
#> $game_settings$league_ko_first_instead_of_random
#> [1] FALSE
#> 
#> $game_settings$cup_start_event_id
#> NULL
#> 
#> $game_settings$cup_stop_event_id
#> NULL
#> 
#> $game_settings$cup_qualifying_method
#> NULL
#> 
#> $game_settings$cup_type
#> NULL
#> 
#> $game_settings$featured_entries
#> list()
#> 
#> $game_settings$element_sell_at_purchase_price
#> [1] FALSE
#> 
#> $game_settings$percentile_ranks
#>  [1]  1  5 10 15 20 25 30 35 40 45 50 55 60 65 70 75 80 95 90 95
#> 
#> $game_settings$underdog_differential
#> [1] 5
#> 
#> $game_settings$squad_squadplay
#> [1] 11
#> 
#> $game_settings$squad_squadsize
#> [1] 15
#> 
#> $game_settings$squad_special_min
#> NULL
#> 
#> $game_settings$squad_special_max
#> NULL
#> 
#> $game_settings$squad_team_limit
#> [1] 3
#> 
#> $game_settings$squad_total_spend
#> [1] 1000
#> 
#> $game_settings$ui_currency_multiplier
#> [1] 10
#> 
#> $game_settings$ui_use_special_shirts
#> [1] FALSE
#> 
#> $game_settings$ui_special_shirt_exclusions
#> list()
#> 
#> $game_settings$stats_form_days
#> [1] 30
#> 
#> $game_settings$sys_vice_captain_enabled
#> [1] TRUE
#> 
#> $game_settings$transfers_cap
#> [1] 20
#> 
#> $game_settings$transfers_sell_on_fee
#> [1] 0.5
#> 
#> $game_settings$max_extra_free_transfers
#> [1] 4
#> 
#> $game_settings$league_h2h_tiebreak_stats
#> [1] "+goals_scored"   "-goals_conceded"
#> 
#> $game_settings$timezone
#> [1] "UTC"
#> 
#> 
#> $game_config
#> $game_config$settings
#> $game_config$settings$entry_per_event
#> [1] FALSE
#> 
#> $game_config$settings$timezone
#> [1] "UTC"
#> 
#> 
#> $game_config$rules
#> $game_config$rules$league_join_private_max
#> [1] 30
#> 
#> $game_config$rules$league_join_public_max
#> [1] 5
#> 
#> $game_config$rules$league_max_size_public_classic
#> [1] 20
#> 
#> $game_config$rules$league_max_size_public_h2h
#> [1] 16
#> 
#> $game_config$rules$league_max_size_private_h2h
#> [1] 16
#> 
#> $game_config$rules$league_max_ko_rounds_private_h2h
#> [1] 3
#> 
#> $game_config$rules$league_prefix_public
#> [1] "League"
#> 
#> $game_config$rules$league_points_h2h_win
#> [1] 3
#> 
#> $game_config$rules$league_points_h2h_lose
#> [1] 0
#> 
#> $game_config$rules$league_points_h2h_draw
#> [1] 1
#> 
#> $game_config$rules$league_ko_first_instead_of_random
#> [1] FALSE
#> 
#> $game_config$rules$cup_start_event_id
#> NULL
#> 
#> $game_config$rules$cup_stop_event_id
#> NULL
#> 
#> $game_config$rules$cup_qualifying_method
#> NULL
#> 
#> $game_config$rules$cup_type
#> NULL
#> 
#> $game_config$rules$featured_entries
#> list()
#> 
#> $game_config$rules$element_sell_at_purchase_price
#> [1] FALSE
#> 
#> $game_config$rules$percentile_ranks
#>  [1]  1  5 10 15 20 25 30 35 40 45 50 55 60 65 70 75 80 95 90 95
#> 
#> $game_config$rules$underdog_differential
#> [1] 5
#> 
#> $game_config$rules$squad_squadplay
#> [1] 11
#> 
#> $game_config$rules$squad_squadsize
#> [1] 15
#> 
#> $game_config$rules$squad_special_min
#> NULL
#> 
#> $game_config$rules$squad_special_max
#> NULL
#> 
#> $game_config$rules$squad_team_limit
#> [1] 3
#> 
#> $game_config$rules$squad_total_spend
#> [1] 1000
#> 
#> $game_config$rules$ui_currency_multiplier
#> [1] 10
#> 
#> $game_config$rules$ui_use_special_shirts
#> [1] FALSE
#> 
#> $game_config$rules$ui_special_shirt_exclusions
#> list()
#> 
#> $game_config$rules$stats_form_days
#> [1] 30
#> 
#> $game_config$rules$sys_vice_captain_enabled
#> [1] TRUE
#> 
#> $game_config$rules$transfers_cap
#> [1] 20
#> 
#> $game_config$rules$transfers_sell_on_fee
#> [1] 0.5
#> 
#> $game_config$rules$max_extra_free_transfers
#> [1] 4
#> 
#> $game_config$rules$league_h2h_tiebreak_stats
#> [1] "+goals_scored"   "-goals_conceded"
#> 
#> 
#> $game_config$scoring
#> $game_config$scoring$long_play
#> [1] 2
#> 
#> $game_config$scoring$short_play
#> [1] 1
#> 
#> $game_config$scoring$goals_conceded
#> $game_config$scoring$goals_conceded$DEF
#> [1] -1
#> 
#> $game_config$scoring$goals_conceded$FWD
#> [1] 0
#> 
#> $game_config$scoring$goals_conceded$GKP
#> [1] -1
#> 
#> $game_config$scoring$goals_conceded$MID
#> [1] 0
#> 
#> $game_config$scoring$goals_conceded$MNG
#> [1] 0
#> 
#> 
#> $game_config$scoring$saves
#> [1] 1
#> 
#> $game_config$scoring$goals_scored
#> $game_config$scoring$goals_scored$DEF
#> [1] 6
#> 
#> $game_config$scoring$goals_scored$FWD
#> [1] 4
#> 
#> $game_config$scoring$goals_scored$GKP
#> [1] 10
#> 
#> $game_config$scoring$goals_scored$MID
#> [1] 5
#> 
#> $game_config$scoring$goals_scored$MNG
#> [1] 0
#> 
#> 
#> $game_config$scoring$assists
#> [1] 3
#> 
#> $game_config$scoring$clean_sheets
#> $game_config$scoring$clean_sheets$DEF
#> [1] 4
#> 
#> $game_config$scoring$clean_sheets$FWD
#> [1] 0
#> 
#> $game_config$scoring$clean_sheets$GKP
#> [1] 4
#> 
#> $game_config$scoring$clean_sheets$MID
#> [1] 1
#> 
#> $game_config$scoring$clean_sheets$MNG
#> [1] 0
#> 
#> 
#> $game_config$scoring$penalties_saved
#> [1] 5
#> 
#> $game_config$scoring$penalties_missed
#> [1] -2
#> 
#> $game_config$scoring$yellow_cards
#> [1] -1
#> 
#> $game_config$scoring$red_cards
#> [1] -3
#> 
#> $game_config$scoring$own_goals
#> [1] -2
#> 
#> $game_config$scoring$bonus
#> [1] 1
#> 
#> $game_config$scoring$bps
#> [1] 0
#> 
#> $game_config$scoring$influence
#> [1] 0
#> 
#> $game_config$scoring$creativity
#> [1] 0
#> 
#> $game_config$scoring$threat
#> [1] 0
#> 
#> $game_config$scoring$ict_index
#> [1] 0
#> 
#> $game_config$scoring$special_multiplier
#> [1] 1
#> 
#> $game_config$scoring$mng_goals_scored
#> $game_config$scoring$mng_goals_scored$DEF
#> [1] 0
#> 
#> $game_config$scoring$mng_goals_scored$FWD
#> [1] 0
#> 
#> $game_config$scoring$mng_goals_scored$GKP
#> [1] 0
#> 
#> $game_config$scoring$mng_goals_scored$MID
#> [1] 0
#> 
#> $game_config$scoring$mng_goals_scored$MNG
#> [1] 1
#> 
#> 
#> $game_config$scoring$mng_clean_sheets
#> $game_config$scoring$mng_clean_sheets$DEF
#> [1] 0
#> 
#> $game_config$scoring$mng_clean_sheets$FWD
#> [1] 0
#> 
#> $game_config$scoring$mng_clean_sheets$GKP
#> [1] 0
#> 
#> $game_config$scoring$mng_clean_sheets$MID
#> [1] 0
#> 
#> $game_config$scoring$mng_clean_sheets$MNG
#> [1] 2
#> 
#> 
#> $game_config$scoring$mng_win
#> $game_config$scoring$mng_win$DEF
#> [1] 0
#> 
#> $game_config$scoring$mng_win$FWD
#> [1] 0
#> 
#> $game_config$scoring$mng_win$GKP
#> [1] 0
#> 
#> $game_config$scoring$mng_win$MID
#> [1] 0
#> 
#> $game_config$scoring$mng_win$MNG
#> [1] 6
#> 
#> 
#> $game_config$scoring$mng_draw
#> $game_config$scoring$mng_draw$DEF
#> [1] 0
#> 
#> $game_config$scoring$mng_draw$FWD
#> [1] 0
#> 
#> $game_config$scoring$mng_draw$GKP
#> [1] 0
#> 
#> $game_config$scoring$mng_draw$MID
#> [1] 0
#> 
#> $game_config$scoring$mng_draw$MNG
#> [1] 3
#> 
#> 
#> $game_config$scoring$mng_loss
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_win
#> $game_config$scoring$mng_underdog_win$DEF
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_win$FWD
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_win$GKP
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_win$MID
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_win$MNG
#> [1] 10
#> 
#> 
#> $game_config$scoring$mng_underdog_draw
#> $game_config$scoring$mng_underdog_draw$DEF
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_draw$FWD
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_draw$GKP
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_draw$MID
#> [1] 0
#> 
#> $game_config$scoring$mng_underdog_draw$MNG
#> [1] 5
#> 
#> 
#> $game_config$scoring$starts
#> [1] 0
#> 
#> $game_config$scoring$expected_assists
#> [1] 0
#> 
#> $game_config$scoring$expected_goal_involvements
#> [1] 0
#> 
#> $game_config$scoring$expected_goals_conceded
#> [1] 0
#> 
#> $game_config$scoring$expected_goals
#> [1] 0
#> 
#> 
#> 
#> $phases
#>    id      name start_event stop_event highest_score
#> 1   1   Overall           1         38          2380
#> 2   2    August           1          3           309
#> 3   3 September           4          6           301
#> 4   4   October           7          9           267
#> 5   5  November          10         13           362
#> 6   6  December          14         19           551
#> 7   7   January          20         23           399
#> 8   8  February          24         27           467
#> 9   9     March          28         29           205
#> 10 10     April          30         34           287
#> 11 11       May          35         38            NA
#> 
#> $teams
#>    code draw form id loss           name played points position short_name
#> 1     3    0   NA  1    0        Arsenal      0      0        2        ARS
#> 2     7    0   NA  2    0    Aston Villa      0      0        7        AVL
#> 3    91    0   NA  3    0    Bournemouth      0      0        8        BOU
#> 4    94    0   NA  4    0      Brentford      0      0       11        BRE
#> 5    36    0   NA  5    0       Brighton      0      0       10        BHA
#> 6     8    0   NA  6    0        Chelsea      0      0        6        CHE
#> 7    31    0   NA  7    0 Crystal Palace      0      0       12        CRY
#> 8    11    0   NA  8    0        Everton      0      0       13        EVE
#> 9    54    0   NA  9    0         Fulham      0      0        9        FUL
#> 10   40    0   NA 10    0        Ipswich      0      0       18        IPS
#> 11   13    0   NA 11    0      Leicester      0      0       19        LEI
#> 12   14    0   NA 12    0      Liverpool      0      0        1        LIV
#> 13   43    0   NA 13    0       Man City      0      0        5        MCI
#> 14    1    0   NA 14    0        Man Utd      0      0       14        MUN
#> 15    4    0   NA 15    0      Newcastle      0      0        3        NEW
#> 16   17    0   NA 16    0  Nott'm Forest      0      0        4        NFO
#> 17   20    0   NA 17    0    Southampton      0      0       20        SOU
#> 18    6    0   NA 18    0          Spurs      0      0       15        TOT
#> 19   21    0   NA 19    0       West Ham      0      0       17        WHU
#> 20   39    0   NA 20    0         Wolves      0      0       16        WOL
#>    strength team_division unavailable win strength_overall_home
#> 1         4            NA       FALSE   0                  1235
#> 2         3            NA       FALSE   0                  1110
#> 3         4            NA       FALSE   0                  1195
#> 4         3            NA       FALSE   0                  1100
#> 5         3            NA       FALSE   0                  1140
#> 6         4            NA       FALSE   0                  1230
#> 7         3            NA       FALSE   0                  1130
#> 8         3            NA       FALSE   0                  1095
#> 9         3            NA       FALSE   0                  1150
#> 10        3            NA       FALSE   0                  1075
#> 11        3            NA       FALSE   0                  1075
#> 12        5            NA       FALSE   0                  1355
#> 13        4            NA       FALSE   0                  1230
#> 14        3            NA       FALSE   0                  1115
#> 15        4            NA       FALSE   0                  1195
#> 16        4            NA       FALSE   0                  1195
#> 17        2            NA       FALSE   0                  1055
#> 18        3            NA       FALSE   0                  1150
#> 19        3            NA       FALSE   0                  1095
#> 20        3            NA       FALSE   0                  1095
#>    strength_overall_away strength_attack_home strength_attack_away
#> 1                   1335                 1220                 1370
#> 2                   1215                 1080                 1160
#> 3                   1215                 1130                 1170
#> 4                   1195                 1080                 1080
#> 5                   1150                 1120                 1140
#> 6                   1235                 1150                 1150
#> 7                   1130                 1140                 1140
#> 8                   1125                 1130                 1140
#> 9                   1150                 1150                 1150
#> 10                  1075                 1070                 1070
#> 11                  1085                 1060                 1080
#> 12                  1360                 1340                 1350
#> 13                  1250                 1140                 1150
#> 14                  1140                 1120                 1120
#> 15                  1225                 1140                 1150
#> 16                  1220                 1200                 1250
#> 17                  1065                 1050                 1050
#> 18                  1185                 1090                 1090
#> 19                  1100                 1080                 1080
#> 20                  1095                 1070                 1070
#>    strength_defence_home strength_defence_away pulse_id
#> 1                   1250                  1300        1
#> 2                   1140                  1270        2
#> 3                   1260                  1260      127
#> 4                   1120                  1310      130
#> 5                   1160                  1160      131
#> 6                   1310                  1320        4
#> 7                   1120                  1120        6
#> 8                   1070                  1100        7
#> 9                   1150                  1150       34
#> 10                  1080                  1080        8
#> 11                  1090                  1090       26
#> 12                  1370                  1370       10
#> 13                  1320                  1350       11
#> 14                  1110                  1160       12
#> 15                  1250                  1300       23
#> 16                  1190                  1190       15
#> 17                  1060                  1080       20
#> 18                  1210                  1280       21
#> 19                  1110                  1120       25
#> 20                  1120                  1120       38
#>